refactor: extract pipeline/lifecycle.rs
Move the lockfile-driven lifecycle commands (remove, doctor, update,
list) out of pipeline/mod.rs into a focused submodule. These four
operations consult the lockfile to know what is installed and call
the leaf modules (output, hash, git) plus the top-level install
coordinator — they own no parsing or render logic themselves.

Public API is preserved via 'pub use lifecycle::{...}' in mod.rs so
existing imports in main.rs keep working unchanged.

Refs #180
