+++ src/pipeline/mod.rs
@@ -33,6 +33,7 @@ use crate::source::InstallSource;
 mod discovery;
 mod git;
 mod hash;
+mod lifecycle;
 mod output;
 mod report;
 
@@ -42,7 +43,7 @@ use discovery::{
 };
 pub use git::{fetch_ssot, install_from_git_url, install_from_source};
 pub(crate) use hash::hash_item_dir;
-use hash::hash_source_items;
+pub use lifecycle::{doctor, list, remove, update};
 use output::{output_path, remove_item_outputs, write_output};
 pub use report::*;
 
@@ -650,440 +651,6 @@ fn install_plugins_from_bundles(
     }
 
     results
-}
-
-/// Remove installed content recorded in `<target>/.upskill-lock.json`,
-/// matching `filter`. For each matching entry, deletes every per-client
-/// output file (per [`output_path`]) and drops the entry from the
-/// lockfile. Best-effort `rmdir` of empty per-item parent directories
-/// (e.g., `.claude/skills/<name>/`) so the workspace stays clean.
-///
-/// Ancillary files (`CLAUDE.md`, `opencode.json`,
-/// `.vscode/settings.json`) are deliberately left alone — they are
-/// user-owned after creation per ADR-0003.
-pub fn remove(target: &Path, filter: RemoveFilter) -> Result<RemoveReport> {
-    let mut lock = crate::lockfile::Lockfile::load(target)?;
-
-    let to_remove: Vec<crate::lockfile::LockedItem> = match &filter {
-        RemoveFilter::ByNames(names) => lock
-            .items
-            .iter()
-            .filter(|i| names.iter().any(|n| n == &i.name))
-            .cloned()
-            .collect(),
-        RemoveFilter::BySource(source) => lock
-            .items
-            .iter()
-            .filter(|i| &i.source == source)
-            .cloned()
-            .collect(),
-    };
-
-    if let RemoveFilter::ByNames(names) = &filter {
-        let matched: std::collections::BTreeSet<&str> =
-            to_remove.iter().map(|i| i.name.as_str()).collect();
-        let unknown: Vec<&str> = names
-            .iter()
-            .filter(|n| !matched.contains(n.as_str()))
-            .map(String::as_str)
-            .collect();
-        if !unknown.is_empty() {
-            anyhow::bail!("not in lockfile: {}", unknown.join(", "));
-        }
-    }
-
-    let mut report = RemoveReport::default();
-    for entry in &to_remove {
-        let kind = entry.kind;
-        let mut deleted_files = Vec::new();
-        for client in ALL_CLIENTS {
-            let rel = output_path(kind, client, &entry.name);
-            let full = target.join(&rel);
-            if full.exists() {
-                fs::remove_file(&full).with_context(|| format!("delete {}", full.display()))?;
-                deleted_files.push(rel);
-                if let Some(parent) = full.parent() {
-                    let _ = fs::remove_dir(parent);
-                }
-            }
-        }
-        lock.remove(entry.kind, &entry.name);
-        report.items.push(RemovedItem {
-            kind,
-            name: entry.name.clone(),
-            deleted_files,
-        });
-    }
-
-    // Clean up plugins associated with removed bundles.
-    let bundles_to_remove: Vec<String> = match &filter {
-        RemoveFilter::ByNames(names) => lock
-            .bundles
-            .iter()
-            .filter(|b| b.items.iter().any(|item| names.contains(item)))
-            .map(|b| b.name.clone())
-            .collect(),
-        RemoveFilter::BySource(source) => lock
-            .bundles
-            .iter()
-            .filter(|b| &b.source == source)
-            .map(|b| b.name.clone())
-            .collect(),
-    };
-
-    for plugin in lock
-        .plugins
-        .iter()
-        .filter(|p| bundles_to_remove.contains(&p.bundle))
-    {
-        if plugin.client == "opencode"
-            && plugin.status == crate::lockfile::PluginInstallStatus::Installed
-        {
-            let _ = crate::ancillary::remove_opencode_plugin_uri(target, &plugin.identifier);
-        }
-    }
-    lock.plugins
-        .retain(|p| !bundles_to_remove.contains(&p.bundle));
-    lock.bundles
-        .retain(|b| !bundles_to_remove.contains(&b.name));
-
-    lock.save(target)?;
-    Ok(report)
-}
-
-/// Verify installed-state consistency against `.upskill-lock.json`.
-/// Three independent buckets per ADR-0004:
-/// - **missing outputs** — file paths the lockfile says should exist
-///   but don't. Reinstall (`upskill add <source>`) fixes it.
-/// - **stale hashes** — for `local:` sources still on disk, the SSOT
-///   item directory hashes to a value that does not match the lockfile.
-///   `upskill update` (or `--dry-run`) fixes it.
-/// - **orphan entries** — the lockfile points at a `local:` source that
-///   is gone, or at an item that no longer exists in its source. The
-///   user has to `upskill remove` to clear it.
-///
-/// Doctor never fetches. Remote-source drift is detected by
-/// `update --dry-run`, which does fetch.
-pub fn doctor(target: &Path) -> Result<DoctorReport> {
-    let lock = crate::lockfile::Lockfile::load(target)?;
-    let mut report = DoctorReport::default();
-
-    for entry in &lock.items {
-        let kind = entry.kind;
-
-        let mut missing = Vec::new();
-        for client in ALL_CLIENTS {
-            let rel = output_path(kind, client, &entry.name);
-            if !target.join(&rel).exists() {
-                missing.push(rel);
-            }
-        }
-        if !missing.is_empty() {
-            report.missing_outputs.push(MissingOutput {
-                kind,
-                name: entry.name.clone(),
-                missing_files: missing,
-            });
-        }
-
-        if let Some(local_path) = entry.source.strip_prefix("local:") {
-            let ssot_root = Path::new(local_path);
-            if !ssot_root.is_dir() {
-                report.orphan_entries.push(OrphanEntry {
-                    kind,
-                    name: entry.name.clone(),
-                    source: entry.source.clone(),
-                    reason: OrphanReason::LocalPathGone,
-                });
-                continue;
-            }
-            let item_dir = ssot_root.join(&entry.name);
-            if !item_dir.is_dir() {
-                report.orphan_entries.push(OrphanEntry {
-                    kind,
-                    name: entry.name.clone(),
-                    source: entry.source.clone(),
-                    reason: OrphanReason::ItemMissingInSource,
-                });
-                continue;
-            }
-            let current = hash_item_dir(&item_dir);
-            if current != entry.hash {
-                report.stale_hashes.push(StaleHash {
-                    kind,
-                    name: entry.name.clone(),
-                    source: entry.source.clone(),
-                    stored_hash: entry.hash.clone(),
-                    current_hash: current,
-                });
-            }
-        }
-        // Non-local sources: doctor only validates per-client outputs.
-        // Hash comparison would require a network fetch — out of scope
-        // here, see `update --dry-run`.
-    }
-
-    // -- Plugin reconciliation (ADR-0008 / issue #151) --
-    // Walk the lockfile's plugin entries and reconcile against each client's
-    // installed plugin list.
-    //
-    // Two buckets:
-    // - skipped_plugins: status == Skipped (CLI was absent at install time).
-    //   Always surface; does not affect is_clean().
-    // - missing_plugins: status == Installed but the client no longer has the
-    //   plugin.  This is drift → is_clean() returns false → exit 1.
-    for plugin in &lock.plugins {
-        use crate::lockfile::PluginInstallStatus;
-        use crate::plugin::{
-            PluginScope, check_claude_plugin_installed, check_opencode_plugin_installed,
-            check_vscode_extension_installed,
-        };
-
-        match &plugin.status {
-            PluginInstallStatus::Skipped | PluginInstallStatus::Instructions => {
-                // Plugin was never installed because the CLI was missing or
-                // requires manual instructions.
-                // Report it so it is not silently ignored.
-                report.skipped_plugins.push(SkippedPlugin {
-                    name: plugin.name.clone(),
-                    client: plugin.client.clone(),
-                    identifier: plugin.identifier.clone(),
-                    bundle: plugin.bundle.clone(),
-                });
-            }
-            PluginInstallStatus::Installed => {
-                // Query the client to verify the plugin is still there.
-                let check = match plugin.client.as_str() {
-                    "claude" => {
-                        let scope = match plugin.scope.as_deref() {
-                            Some("user") => PluginScope::User,
-                            _ => PluginScope::Project,
-                        };
-                        check_claude_plugin_installed(&plugin.name, scope)
-                    }
-                    "vscode" => check_vscode_extension_installed(&plugin.identifier),
-                    "opencode" => check_opencode_plugin_installed(&plugin.identifier),
-                    // Unknown client — skip silently.
-                    _ => continue,
-                };
-                if check.is_not_installed() {
-                    report.missing_plugins.push(MissingPlugin {
-                        name: plugin.name.clone(),
-                        client: plugin.client.clone(),
-                        identifier: plugin.identifier.clone(),
-                        bundle: plugin.bundle.clone(),
-                    });
-                }
-                // CliNotFound or QueryFailed: cannot determine state — skip
-                // silently to avoid false positives.
-            }
-        }
-    }
-
-    Ok(report)
-}
-
-/// Re-fetch every source recorded in `<target>/.upskill-lock.json` and
-/// either reinstall (`Apply`) or report what would change (`DryRun`).
-///
-/// `names` selects which lockfile entries to update; empty means "all".
-/// When names are given, the entries' source labels are used to fetch
-/// — so `update foo` may also re-render other items installed from the
-/// same source (the pipeline always reinstalls a source as a unit).
-/// Names that match no lockfile entry are an error.
-///
-/// `update` always fetches per ADR-0004 — there is no `--offline`.
-pub fn update(
-    target: &Path,
-    names: &[String],
-    mode: UpdateMode,
-    plugin_scope: crate::plugin::PluginScope,
-) -> Result<UpdateReport> {
-    let lock = crate::lockfile::Lockfile::load(target)?;
-
-    let entries: Vec<crate::lockfile::LockedItem> = if names.is_empty() {
-        lock.items.clone()
-    } else {
-        let matched: Vec<crate::lockfile::LockedItem> = lock
-            .items
-            .iter()
-            .filter(|i| names.iter().any(|n| n == &i.name))
-            .cloned()
-            .collect();
-        let matched_names: std::collections::BTreeSet<&str> =
-            matched.iter().map(|i| i.name.as_str()).collect();
-        let unknown: Vec<&str> = names
-            .iter()
-            .filter(|n| !matched_names.contains(n.as_str()))
-            .map(String::as_str)
-            .collect();
-        if !unknown.is_empty() {
-            anyhow::bail!("not in lockfile: {}", unknown.join(", "));
-        }
-        matched
-    };
-
-    // Group by source: installing or hashing once per source covers every
-    // matching entry sharing it.
-    let mut by_source: std::collections::BTreeMap<String, Vec<crate::lockfile::LockedItem>> =
-        std::collections::BTreeMap::new();
-    for entry in entries {
-        by_source
-            .entry(entry.source.clone())
-            .or_default()
-            .push(entry);
-    }
-
-    let mut report = UpdateReport::default();
-    for (source_label, source_entries) in by_source {
-        let source = crate::source::parse_install_source_label(&source_label)
-            .with_context(|| format!("parse lockfile source label `{source_label}`"))?;
-
-        match mode {
-            UpdateMode::Apply => {
-                // Build aliases from lockfile entries that have source_name
-                let aliases: Vec<(String, String)> = source_entries
-                    .iter()
-                    .filter_map(|e| {
-                        e.source_name
-                            .as_ref()
-                            .map(|sn| (sn.clone(), e.name.clone()))
-                    })
-                    .collect();
-                let options = AddOptions {
-                    force: true,
-                    aliases,
-                    excludes: vec![],
-                };
-                let install_report =
-                    install_with_lockfile(&source, target, &[], plugin_scope, &options)?;
-                let mut new_hashes: std::collections::BTreeMap<(ItemKind, String), Option<String>> =
-                    std::collections::BTreeMap::new();
-                for it in &install_report.items {
-                    new_hashes.insert((it.kind, it.name.clone()), it.source_hash.clone());
-                }
-                // Collect orphans first, then batch-remove from lockfile.
-                let mut orphans: Vec<(&crate::lockfile::LockedItem, ItemKind)> = Vec::new();
-                for entry in &source_entries {
-                    let kind = entry.kind;
-                    if !new_hashes.contains_key(&(kind, entry.name.clone())) {
-                        orphans.push((entry, kind));
-                    }
-                }
-                if !orphans.is_empty() {
-                    let mut lock = crate::lockfile::Lockfile::load(target)?;
-                    for (entry, kind) in &orphans {
-                        remove_item_outputs(target, *kind, &entry.name);
-                        lock.remove(entry.kind, &entry.name);
-                        report.items.push(UpdatedItem {
-                            kind: *kind,
-                            name: entry.name.clone(),
-                            source: source_label.clone(),
-                            status: UpdateStatus::Removed,
-                        });
-                    }
-                    lock.save(target)?;
-                }
-                for entry in &source_entries {
-                    let kind = entry.kind;
-                    if !new_hashes.contains_key(&(kind, entry.name.clone())) {
-                        continue; // already handled as orphan
-                    }
-                    let new_hash = new_hashes
-                        .get(&(kind, entry.name.clone()))
-                        .cloned()
-                        .flatten();
-                    let status = if new_hash == entry.hash {
-                        UpdateStatus::UpToDate
-                    } else {
-                        UpdateStatus::Updated {
-                            old_hash: entry.hash.clone(),
-                            new_hash,
-                        }
-                    };
-                    report.items.push(UpdatedItem {
-                        kind,
-                        name: entry.name.clone(),
-                        source: source_label.clone(),
-                        status,
-                    });
-                }
-            }
-            UpdateMode::DryRun => {
-                let (root, _guard) = fetch_ssot(&source)?;
-                let new_hashes = hash_source_items(&root);
-                for entry in &source_entries {
-                    let kind = entry.kind;
-                    let lookup_name = entry.source_name.as_deref().unwrap_or(&entry.name);
-                    if !new_hashes.contains_key(&(kind, lookup_name.to_string())) {
-                        report.items.push(UpdatedItem {
-                            kind,
-                            name: entry.name.clone(),
-                            source: source_label.clone(),
-                            status: UpdateStatus::WouldRemove,
-                        });
-                        continue;
-                    }
-                    let new_hash = new_hashes
-                        .get(&(kind, lookup_name.to_string()))
-                        .cloned()
-                        .flatten();
-                    let status = if new_hash == entry.hash {
-                        UpdateStatus::UpToDate
-                    } else {
-                        UpdateStatus::WouldChange {
-                            old_hash: entry.hash.clone(),
-                            new_hash,
-                        }
-                    };
-                    report.items.push(UpdatedItem {
-                        kind,
-                        name: entry.name.clone(),
-                        source: source_label.clone(),
-                        status,
-                    });
-                }
-            }
-        }
-    }
-
-    Ok(report)
-}
-
-/// List installed content from `<target>/.upskill-lock.json`. No
-/// filesystem walk, no fetch — pure lockfile dump grouped by kind.
-/// Empty lockfile (or missing file) is not an error; the returned
-/// report is empty.
-pub fn list(target: &Path) -> Result<ListReport> {
-    let lock = crate::lockfile::Lockfile::load(target)?;
-    let mut report = ListReport::default();
-    for entry in &lock.items {
-        let kind = entry.kind;
-        let listed = ListedItem {
-            kind,
-            name: entry.name.clone(),
-            source: entry.source.clone(),
-            git_ref: entry.git_ref.clone(),
-        };
-        match kind {
-            ItemKind::Rule => report.rules.push(listed),
-            ItemKind::Skill => report.skills.push(listed),
-            ItemKind::Agent => report.agents.push(listed),
-        }
-    }
-    for bucket in [&mut report.rules, &mut report.skills, &mut report.agents] {
-        bucket.sort_by(|a, b| a.name.cmp(&b.name));
-    }
-    for bundle in &lock.bundles {
-        report.bundles.push(ListedBundle {
-            name: bundle.name.clone(),
-            source: bundle.source.clone(),
-            git_ref: bundle.git_ref.clone(),
-            items: bundle.items.clone(),
-        });
-    }
-    report.bundles.sort_by(|a, b| a.name.cmp(&b.name));
-    Ok(report)
 }
 
 fn install_items_of_kind(
